+++ minicbor/src/lib.rs
@@ -140,9 +140,6 @@
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
-#[cfg(feature = "legacy")]
-pub mod legacy;
-
 pub mod bytes;
 pub mod data;
 pub mod decode;
@@ -167,18 +164,35 @@ pub use minicbor_derive::*;
 /// Decode a type implementing [`Decode`] from the given byte slice.
 pub fn decode<'b, T>(b: &'b [u8]) -> Result<T, decode::Error>
 where
-    T: Decode<'b>
+    T: Decode<'b, ()>
 {
     Decoder::new(b).decode()
+}
+
+/// Decode a type implementing [`Decode`] from the given byte slice.
+pub fn decode_with<'b, C, T>(b: &'b [u8], ctx: &mut C) -> Result<T, decode::Error>
+where
+    T: Decode<'b, C>
+{
+    Decoder::new(b).decode_with(ctx)
 }
 
 /// Encode a type implementing [`Encode`] to the given [`encode::Write`] impl.
 pub fn encode<T, W>(x: T, w: W) -> Result<(), encode::Error<W::Error>>
 where
-    T: Encode,
+    T: Encode<()>,
     W: encode::Write
 {
     Encoder::new(w).encode(x)?.ok()
+}
+
+/// Encode a type implementing [`Encode`] to the given [`encode::Write`] impl.
+pub fn encode_with<C, T, W>(x: T, w: W, ctx: &mut C) -> Result<(), encode::Error<W::Error>>
+where
+    T: Encode<C>,
+    W: encode::Write
+{
+    Encoder::new(w).encode_with(x, ctx)?.ok()
 }
 
 /// Encode a type implementing [`Encode`] and return the encoded byte vector.
@@ -187,11 +201,24 @@ where
 #[cfg(feature = "std")]
 pub fn to_vec<T>(x: T) -> Result<Vec<u8>, encode::Error<std::io::Error>>
 where
-    T: Encode
+    T: Encode<()>
 {
     let mut e = Encoder::new(Vec::new());
-    x.encode(&mut e)?;
-    Ok(e.into_inner())
+    x.encode(&mut e, &mut ())?;
+    Ok(e.into_writer())
+}
+
+/// Encode a type implementing [`Encode`] and return the encoded byte vector.
+///
+/// *Requires feature* `"std"`.
+#[cfg(feature = "std")]
+pub fn to_vec_with<C, T>(x: T, ctx: &mut C) -> Result<Vec<u8>, encode::Error<std::io::Error>>
+where
+    T: Encode<C>
+{
+    let mut e = Encoder::new(Vec::new());
+    x.encode(&mut e, ctx)?;
+    Ok(e.into_writer())
 }
 
 /// Display the given CBOR bytes in [diagnostic notation][1].
@@ -227,4 +254,8 @@ where
 pub fn display<'b>(cbor: &'b [u8]) -> impl core::fmt::Display + 'b {
     decode::Tokenizer::new(cbor)
 }
+
+// Ensure we can safely cast a `usize` to a `u64`.
+const __USIZE_FITS_INTO_U64: () =
+    assert!(core::mem::size_of::<usize>() <= core::mem::size_of::<u64>());